added global activity animation, share and rate methods and cleaned some other stuff
when added the global animation to the style file we don't need to give animation to the intent when starting and closing it and so when wont need onBackPressed anymore.
share and rate the app has been added to more menu.
call us activity has been created just for the access and preview but hasn't been designed yet.
the backPressed boolean in splash has been deleted becase we are using handler.removecallbacks and messages instead which is the right way.